Add foundations for svg support *
Using xml.etree.ElementTree to extract data from an svg file
Create SVG class to represent an svg file and a Rect class to
represent a rect element in an svg file with a draw function that
draws itself onto a specified pygame.Surface object. Reorganised
code with a menu_screen function that draws the menu screen.
The Sprite class now has an image parameter in instantiation that
is used as it's "image" attribute and supplies the Sprite's "size"
attribute "size" instantiation parameter has been replaced with
"pos".